Moving the Logic to a new Directory
