+++ modules/getChat.js
@@ -1,8 +1,8 @@
 //Libary Imports
 const request = require('request-promise'); //The Request Module for sending the different Modules
-const endpoints = require('../endpoints.js'); //For Creating shorter URL's in this Module
-const objs = require('../objects.js'); //For Storing the Objects that the Framework returns. 
-const sorter = require('../sorter.js'); //For easier Sorting of various Responses.
+const endpoints = require('../helpers/endpoints.js'); //For Creating shorter URL's in this Module
+const objs = require('../helpers/objects.js'); //For Storing the Objects that the Framework returns. 
+const sorter = require('../helpers/sorter.js'); //For easier Sorting of various Responses.
 const { getConfig } = require('../index');
 
 /**
