+++ modules/login.js
@@ -1,6 +1,6 @@
 //Libary Imports
 const request = require('request-promise'); //The Request Module for sending the different Modules
-const endpoints = require('../endpoints.js'); //For Creating shorter URL's in this Module
+const endpoints = require('../helpers/endpoints.js'); //For Creating shorter URL's in this Module
 const { setConfig } = require('../index');
 
 /**
